refactor: MvcController와 LegacyController 분리
